Refactor structure move Calculator.ts to a new folder
